# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GlobeTrotter API
  description: |
    A production-ready, horizontally scalable backend for the GlobeTrotter travel planning system.

    ## Features
    - JWT-based authentication with access tokens
    - User management with profile customization
    - Trip planning and sharing capabilities
    - File upload for user avatars
    - Rate limiting and security middleware
    - Comprehensive error handling

    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
      Authorization: Bearer <your_access_token>
  version: 1.0.0
  contact:
    name: GlobeTrotter Development Team
    email: dev@globetrotter.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.globetrotter.com/api/v1
    description: Production server
security:
  - bearerAuth: []
paths:
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account with email and password authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
            example:
              name: "John Doe"
              email: "john@example.com"
              password: "securepassword123"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthResponse"
              example:
                success: true
                message: "User created successfully"
                data:
                  access_token: "eyJhbGciOiJIUzI1NiIs..."
                  refresh_token: "eyJhbGciOiJIUzI1NiIs..."
                  user:
                    id: "507f1f77bcf86cd799439011"
                    name: "John Doe"
                    email: "john@example.com"
                    avatar_path: ""
                    preferences:
                      language: "en"
                      currency: "USD"
                      theme: "light"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: Authenticates a user with email and password, returns JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: "john@example.com"
              password: "securepassword123"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          description: MongoDB ObjectID
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        avatar_path:
          type: string
          example: "/uploads/profile_pics/avatar_123.jpg"
        preferences:
          $ref: "#/components/schemas/UserPreferences"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserPreferences:
      type: object
      properties:
        language:
          type: string
          example: "en"
        currency:
          type: string
          example: "USD"
        theme:
          type: string
          example: "light"
...
    SignupRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          minLength: 8
          example: "securepassword123"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          example: "securepassword123"
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIs..."

        user:
          $ref: "#/components/schemas/User"
...
    APIResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        error:
          type: string
...
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            success: false
            error: "Invalid request data"
    Unauthorized:
      description: Unauthorized - authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            success: false
            error: "Authentication required"
...
    Conflict:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            success: false
            error: "Email already exists"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            success: false
            error: "Internal server error"
...